# backend/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Centoire API",
    "version": "0.1.0",
    "description": "Centoire community platform API"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string"
          }
        },
        "required": [
          "detail"
        ]
      },
      "TagRef": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ]
      },
      "PublicUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "handle": {
            "type": "string",
            "nullable": true
          },
          "displayName": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "bio": {
            "type": "string",
            "nullable": true
          },
          "reputation": {
            "type": "number"
          },
          "followerCount": {
            "type": "number"
          },
          "followingCount": {
            "type": "number"
          },
          "postCount": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "handle",
          "displayName",
          "avatarUrl",
          "bio",
          "reputation",
          "followerCount",
          "followingCount",
          "postCount",
          "createdAt"
        ]
      },
      "CurrentUser": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PublicUser"
          },
          {
            "type": "object",
            "properties": {
              "email": {
                "type": "string"
              },
              "role": {
                "type": "string",
                "enum": [
                  "member",
                  "admin"
                ]
              },
              "emailVerified": {
                "type": "boolean"
              },
              "onboardingCompleted": {
                "type": "boolean"
              },
              "interests": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/TagRef"
                }
              }
            },
            "required": [
              "email",
              "role",
              "emailVerified",
              "onboardingCompleted",
              "interests"
            ]
          }
        ]
      },
...
      "GoogleLoginInput": {
        "type": "object",
        "properties": {
          "idToken": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "idToken"
        ]
      },
      "VerifyEmailInput": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "token"
        ]
      },
      "ForgotPasswordInput": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordInput": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128
          }
        },
        "required": [
          "token",
          "password"
        ]
      },
      "Sent": {
        "type": "object",
        "properties": {
          "sent": {
            "type": "boolean"
          }
        },
        "required": [
          "sent"
        ]
      },
...
      "Tag": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "style",
              "craft",
              "business",
              "culture"
            ]
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "postCount": {
            "type": "number"
          },
          "followerCount": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "category",
          "description",
          "postCount",
          "followerCount"
        ]
      },
      "Circle": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "about": {
            "type": "string",
            "nullable": true
          },
          "rules": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "coverImageUrl": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "slug": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "name",
                "slug"
              ]
            }
          },
          "memberCount": {
            "type": "number"
          },
          "postCount": {
            "type": "number"
          },
          "viewerRole": {
            "type": "string",
            "nullable": true,
            "enum": [
              "owner",
              "moderator",
              "member"
            ]
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "description",
          "about",
          "rules",
          "avatarUrl",
          "coverImageUrl",
          "tags",
          "memberCount",
          "postCount",
          "viewerRole",
          "createdAt"
        ]
      },
...
      "CreateCircleInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 3,
            "maxLength": 60
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160
          },
          "about": {
            "type": "string",
            "maxLength": 4000
          },
          "rules": {
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 300
            },
            "maxItems": 10
          },
          "tagIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "maxItems": 5
          },
          "avatarUrl": {
            "type": "string",
            "format": "uri"
          },
          "coverImageUrl": {
            "type": "string",
            "format": "uri"
          }
        },
        "required": [
          "name",
          "description"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/auth/google": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "loginWithGoogle",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleLoginInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged in; session cookie set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUser"
                }
              }
            }
          },
          "401": {
            "description": "Google token invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/verify-email": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "verifyEmail",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email verified",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUser"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/forgot-password": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "forgotPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordInput"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Reset email queued if the account exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sent"
                }
              }
            }
          }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "resetPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password updated; session cookie set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUser"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/tags": {
      "get": {
        "tags": [
          "tags"
        ],
        "operationId": "listTags",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "style",
                "craft",
                "business",
                "culture"
              ]
            },
            "required": false,
            "name": "category",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "All tags",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tag"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/circles": {
      "get": {
        "tags": [
          "circles"
        ],
        "operationId": "listCircles",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "q",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "tag",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Circle directory",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Circle"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "circles"
        ],
        "operationId": "createCircle",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCircleInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Circle created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Circle"
                }
              }
            }
          }
        }
      }
    },
...
    "/circles/{slug}/join": {
      "post": {
        "tags": [
          "circles"
        ],
        "operationId": "joinCircle",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "slug",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Joined"
          },
          "404": {
            "description": "Circle not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "circles"
        ],
        "operationId": "leaveCircle",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "slug",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Left"
          },
          "422": {
            "description": "Owners cannot leave",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}